Add cid.[ch] for containing connection ID stuffs
This patch adds cid.h and cid.c to contain connection ID related stuffs
to prevent quicly.c explosion.
